refactor: move hook.utils to top-level utils
