+++ app/src/main/java/io/github/twyora/douyinenhancer/utils/HookTransaction.kt
@@ -1,4 +1,4 @@
-package io.github.twyora.douyinenhancer.hook.utils
+package io.github.twyora.douyinenhancer.utils
 
 import com.highcapable.yukihookapi.hook.core.YukiMemberHookCreator.MemberHookCreator
 import com.highcapable.yukihookapi.hook.log.YLog
